Added filtering of scheduled transactions Computed field
Adding an option to the user client to filter scheduled transaction from
bankrs to include or exclude transactions based on the state of the
Computed field.

Computed transactions are true, if they originate from a Standing order -
the nearest occurrence in the future. The field is false, if the txs is
a regular tx scheduled explicitly for a time in future.